MT 37205: Log planning actions and ability to undo/redo

+++ public/planning/poste/ajax.updateCell.php
@@ -15,6 +15,7 @@ Cette page est appelée par la function JavaScript "bataille_navale" utilisé pa
 */
 
 use App\Model\Position;
+use App\Model\PlanningPositionHistory;
 use App\PlanningBiblio\Helper\PlanningPositionHistoryHelper;
 
 ini_set("display_errors", 0);
@@ -33,7 +34,7 @@ require_once __DIR__."/../../init_ajax.php";
 
 //	Initialisation des variables
 $ajouter=filter_input(INPUT_POST, "ajouter", FILTER_CALLBACK, array("options"=>"sanitize_on"));
-$barrer=filter_input(INPUT_POST, "barrer", FILTER_CALLBACK, array("options"=>"sanitize_on"));
+$barrer=filter_input(INPUT_POST, "barrer", FILTER_SANITIZE_NUMBER_INT);
 $CSRFToken=filter_input(INPUT_POST, "CSRFToken", FILTER_SANITIZE_STRING);
 $date=filter_input(INPUT_POST, "date", FILTER_CALLBACK, array("options"=>"sanitize_dateSQL"));
 $debut=filter_input(INPUT_POST, "debut", FILTER_CALLBACK, array("options"=>"sanitize_time"));
@@ -44,6 +45,7 @@ $perso_id_origine=filter_input(INPUT_POST, "perso_id_origine", FILTER_SANITIZE_N
 $poste=filter_input(INPUT_POST, "poste", FILTER_SANITIZE_NUMBER_INT);
 $site=filter_input(INPUT_POST, "site", FILTER_SANITIZE_NUMBER_INT);
 $tout=filter_input(INPUT_POST, "tout", FILTER_CALLBACK, array("options"=>"sanitize_on"));
+$logaction=filter_input(INPUT_POST, "logaction", FILTER_CALLBACK, array("options"=>"sanitize_on"));
 
 $login_id=$_SESSION['login_id'];
 $now=date("Y-m-d H:i:s");
@@ -53,11 +55,13 @@ $now=date("Y-m-d H:i:s");
 // Suppression ou marquage absent
 if (is_numeric($perso_id) and $perso_id == 0) {
     // Tout barrer
-    if ($barrer and $tout) {
+    if ($barrer == 1 and $tout) {
 
         // History
-        $history = new PlanningPositionHistoryHelper();
-        $history->cross($date, $debut, $fin, $site, $poste, $login_id);
+        if ($logaction) {
+            $history = new PlanningPositionHistoryHelper();
+            $history->cross($date, $debut, $fin, $site, $poste, $login_id);
+        }
 
         $set=array("absent"=>"1", "chgt_login"=>$login_id, "chgt_time"=>$now);
         $where=array("date"=>$date, "debut"=>$debut, "fin"=>$fin, "poste"=>$poste, "site"=>$site);
@@ -66,13 +70,21 @@ if (is_numeric($perso_id) and $perso_id == 0) {
         $db->update("pl_poste", $set, $where);
 
     // Barrer l'agent sélectionné
-    } elseif ($barrer) {
+    } elseif ($barrer == 1) {
 
         // History
-        $history = new PlanningPositionHistoryHelper();
-        $history->cross($date, $debut, $fin, $site, $poste, $login_id, $perso_id_origine);
+        if ($logaction) {
+            $history = new PlanningPositionHistoryHelper();
+            $history->cross($date, $debut, $fin, $site, $poste, $login_id, $perso_id_origine);
+        }
 
         $set=array("absent"=>"1", "chgt_login"=>$login_id, "chgt_time"=>$now);
+        $where=array("date"=>$date, "debut"=>$debut, "fin"=>$fin, "poste"=>$poste, "site"=>$site, "perso_id"=>$perso_id_origine);
+        $db=new db();
+        $db->CSRFToken = $CSRFToken;
+        $db->update("pl_poste", $set, $where);
+    } elseif ($barrer == -1) {
+        $set=array("absent"=>"0", "chgt_login"=>$login_id, "chgt_time"=>$now);
         $where=array("date"=>$date, "debut"=>$debut, "fin"=>$fin, "poste"=>$poste, "site"=>$site, "perso_id"=>$perso_id_origine);
         $db=new db();
         $db->CSRFToken = $CSRFToken;
@@ -82,19 +94,24 @@ if (is_numeric($perso_id) and $perso_id == 0) {
     elseif ($tout) {
 
         // History
-        $history = new PlanningPositionHistoryHelper();
-        $history->delete($date, $debut, $fin, $site, $poste, $login_id);
+        if ($logaction) {
+            $history = new PlanningPositionHistoryHelper();
+            $history->delete($date, $debut, $fin, $site, $poste, $login_id);
+        }
 
         $where=array("date"=>$date, "debut"=>$debut, "fin"=>$fin, "poste"=>$poste, "site"=>$site);
         $db=new db();
         $db->CSRFToken = $CSRFToken;
         $db->delete("pl_poste", $where);
     // Supprimer l'agent sélectionné
-    } else {
+    // FIXME à vérifier. Pas de suppression si on dégrise.
+    } elseif ($griser != -1) {
 
         // History
-        $history = new PlanningPositionHistoryHelper();
-        $history->delete($date, $debut, $fin, $site, $poste, $login_id, $perso_id_origine);
+        if ($logaction) {
+            $history = new PlanningPositionHistoryHelper();
+            $history->delete($date, $debut, $fin, $site, $poste, $login_id, $perso_id_origine);
+        }
 
         $where=array("date"=>$date, "debut"=>$debut, "fin"=>$fin, "poste"=>$poste, "site"=>$site, "perso_id"=>$perso_id_origine);
         $db=new db();
@@ -105,7 +122,14 @@ if (is_numeric($perso_id) and $perso_id == 0) {
 // Remplacement
 else {
     // si ni barrer, ni ajouter : on remplace
-    if (!$barrer and !$ajouter) {
+    if ($barrer == 0 and !$ajouter) {
+
+        // History
+        if ($logaction) {
+            $history = new PlanningPositionHistoryHelper();
+            $history->put($date, $debut, $fin, $site, $poste, $login_id, $perso_id);
+        }
+
         // Suppression des anciens éléments
         $where=array("date"=>$date, "debut"=>$debut, "fin"=>$fin, "poste"=>$poste, "site"=>$site, "perso_id"=> $perso_id_origine);
         $db=new db();
@@ -128,7 +152,7 @@ else {
         }
     }
     // Si barrer : on barre l'ancien et ajoute le nouveau
-    elseif ($barrer) {
+    elseif ($barrer == 1) {
         // On barre l'ancien
         $set=array("absent"=>"1", "chgt_login"=>$login_id, "chgt_time"=>$now);
         $where=array("date"=>$date, "debut"=>$debut, "fin"=>$fin, "poste"=>$poste, "site"=>$site, "perso_id"=>$perso_id_origine);
@@ -155,6 +179,13 @@ else {
 
 // Griser les cellule
 if ($griser == 1) {
+
+    // History
+    if ($logaction) {
+        $history = new PlanningPositionHistoryHelper();
+        $history->disable($date, $debut, $fin, $site, $poste, $login_id, $perso_id_origine);
+    }
+
     $insert=array("date"=>$date, "debut"=>$debut, "fin"=>$fin, "poste"=>$poste, "site"=>$site, "perso_id"=>'0', "grise"=>'1', "chgt_login"=>$login_id, "chgt_time"=>$now);
     $db=new db();
     $db->CSRFToken = $CSRFToken;
@@ -180,13 +211,35 @@ $db->selectLeftJoin(
   "ORDER BY nom,prenom"
 );
 
+$response = array(
+    'tab' => null,
+    'undoable' => 1,
+    'redoable' => 1
+);
+
+$undoables = $entityManager
+     ->getRepository(PlanningPositionHistory::class)
+     ->undoable($date, $site);
+$redoables = $entityManager
+     ->getRepository(PlanningPositionHistory::class)
+     ->redoable($date, $site);
+
+if (empty($undoables)) {
+    $response['undoable'] = 0;
+}
+
+if (empty($redoables)) {
+    $response['redoable'] = 0;
+}
+
 if (!$db->result) {
-    echo json_encode(array());
+    echo json_encode($response);
     return;
 }
 
 if ($db->result[0]['grise'] == 1) {
-    echo json_encode("grise");
+    $response['tab'] = 'grise';
+    echo json_encode($response);
     return;
 }
 
@@ -272,7 +325,9 @@ if ($config['Conges-Enable']) {
     include "../../conges/ajax.planning.updateCell.php";
 }
 
-echo json_encode($tab);
+$response['tab'] = $tab;
+
+echo json_encode($response);
 
 /*
 Résultat :
